Add a logarithmic implementation of at
After trying it by hand for a while, I gave up and used one of the
techniques described in https://ldionne.com/2015/11/29/efficient-parameter-pack-indexing/
for getting the nth element of a parameter pack. The trick uses pack
expansions and std::index_sequences, which probably use a constant-time
compiler intrinsic. The difference in compile-time is significant: we
went from ~15s to get the last element of void4096 (and needing to
increase the compiler's template recursion limit) to ~5s to get the last
element of void65536 (with the default recursion limit).

Other minor additions were adding the `size` static constexpr member of
type_list, which I think is fine as a member rather than an external
"trait-style" struct. Also added were some first/last convenience types.