Add map operation on type_lists
This takes a type_list and a "type adaptor" – a type that takes a single
template type parameter and contains a type alias `type`. The result is
a type_list of the resulting `::type` for each element of the original
type_list.

For example, we can apply the `std::make_signed` adaptor to a type_list
of unsigned integer types to produce the type_list of the corresponding
signed integer types.