Add a contains binary predicate on type_list
This predicate takes a type_list and a type. The resulting `value` is
`true` if the type is in the given type_list, or `false` otherwise.

Seeing as this computes a value rather than a type, I've also added it
as a member of type_list for convenience. (Member types require the
`typename` keyword, so they aren't necessarily worth it compared to
freestanding types.) This is nicer than remembering the ordering of
parameters to the freestanding `contains` predicate.