Add a for_each function to the type_list
The first foray into actual runtime computation. This adds a static
member function on type_list that takes a single object as a parameter.
The function then invokes that object's templated call operator once for
each element of the type_list. Modelling it this way matches how C++
represents templated lambda functions: as objects of a non-template
class with a templated call operator.

Once we unpack the parameter pack over the lambda invocations, we need
to guarantee the order that they will be executed in. To do so, we use
the unary + fold expression, which evaluates the expressions in order.